+++ packages/react/src/themes/teams/components/Text/textStyles.ts
@@ -13,7 +13,6 @@ export default {
       color,
       important,
       timestamp,
-      animation,
       truncated,
       disabled,
       error,
@@ -28,8 +27,6 @@ export default {
     const colors = v.colorScheme[getColorSchemeKey(color)]
     return {
       ...(color && { color: colors.foreground }),
-      // animations are not working with span, unless display is set to 'inline-block'
-      ...(animation && as === 'span' && { display: 'inline-block' }),
       ...(atMention === true && { color: v.atMentionOtherColor }),
       ...(truncated && { overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }),
       ...(disabled && { color: v.disabledColor }),